backend/services/report-lambda/requirements.txt
PyJWT==2.8.0
boto3
